pyproject.toml
[project]
name = "hass-lyngdorf"
version = "0.3.0"
description = "Home Assistant integration for Lyngdorf audio processors"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-homeassistant-custom-component>=0.13",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "F405",   # * imports
    "ARG001", # unused function argument (common in HA)
    "ARG002", # unused method argument (common in HA)
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
